Allow to read-out huge memory allocations.
When running the example in the top directory PROLEAD is crashing by throwing a std::out_of_range.
While clearly something is wrong with the amount of allocated data (21475178120 kB on my WSL2 based system)
it can be prevented from crashing by reading out the allocated memory in an unsigned 64 bits value.

This is what his patch does.

+++ inc/Hardware/Print.hpp
@@ -25,7 +25,7 @@ namespace Hardware
 		* @brief Prints the current memory consumption. 
 		* @author Nicolai Müller
 		*/		
-        unsigned int MemoryConsumption();
+        uint64_t MemoryConsumption();
 		
 		/**
 		* @brief Prints the command line parameters.
